mutation addCarne agora gera os boletos das parcelas de acordo com input

+++ packages/server/src/graphql/resources/carne/carne.schema.ts
@@ -11,7 +11,7 @@ export const carneTypes = `
     descricao: String!
 
     parcelas: Int
-    valorParcelas: Int
+    valorParcelas: Float
     primeiroVencimento: String
   }
 `;
